+++ R/lockfile.R
@@ -1,3 +1,9 @@
 #nocov start
-lockfile <- function(...) asNamespace("renv")$lockfile(...)
+get_lockfile <- function(...) {
+  if (packageVersion("renv") >= "1.0.0") {
+    renv::lockfile_read(...)
+  } else {
+    asNamespace("renv")$lockfile(...)$data()
+  }
+}
 #nocov enc