+++ R/vise-package.R
@@ -2,7 +2,7 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom utils capture.output getParseData
+#' @importFrom utils capture.output getParseData head packageVersion
 #' @import renv
 ## usethis namespace: end
 NULL